+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Updates between October 25th, 2018 and []
 ## Cloud
 1. Fixed a bug with the `name` attribute that returned the entire directory instead of just the filename (without extension)
 2. Added `.subtract` function, this allows a user to provide their own (properly referenced) DEM for use in normalizing the parent cloud object.
+3. `Cloud.clip` now resets the index of the points dataframe `Cloud.data.points`
+4. Resolved issues instantiating `Cloud` objects using `.laz` files.
 
 ## CloudData
 1. Added a check for empty dataframe before writing to file.
@@ -12,6 +14,16 @@ Updates between October 25th, 2018 and []
 ## Collection
 1. Added `bounding_box` attribute that retrieves the bounding box of the entire collection, used in retiling.
 2. Added `retile` function to split large acquisitions into smaller tile sizes, for now this just splits into quadrants.
+3. Added spatial indexing functionality via `laxpy`, tiles can be clipped without loading the entire file into memory.
+
+## Rasterizer
+1. `Grid` now computes bins starting from the top left of an input `Cloud`. This fixes a lot of unnecessary array flipping further downstream.
+
+## Environment
+1. Added `lastools` to the dependency stack
+
+## Testing Suite
+1. Now testing multiple file types, `.ply`, `.laz` and `.las`
 
 # 0.3.1